requirements.txt
azure-storage==0.36.0
azure-servicebus==0.21.1
pip==9.0.1
numpy==1.13.3
thriftpy==0.3.9
thrift==0.9.3
opencv-python==3.3.0.10
prometheus-client==0.0.21
Pillow==4.0.0
tensorflow==1.4.0
tensorflow-gpu==1.2.1
mock==2.0.0
nose==1.3.7
bnr_python_thrift==0.0.21
semantic_version
gitpython
redis==2.10.6
scipy==0.13.3

